# Backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Le Vieux Grimoire API",
        "version": "1.0.0",
        "description": "API documentation for Le Vieux Grimoire website"
    },
    "servers": [
        {
            "url": "http://localhost:4000/api/"
        }
    ],
    "paths": {
        "/auth/signup": {
            "post": {
                "summary": "Hachage du mot de passe de l'utilisateur, ajout de l'utilisateur à la base de données.",
                "tags": [
                    "Authentication"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "description": "User's email address"
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "User's password"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "400": {
                        "description": "Invalid request data"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "summary": "Vérification des informations d'identification de l'utilisateur ; renvoie l’_id de l'utilisateur depuis la base de données et un token web JSON signé (contenant également l'_id de l'utilisateur).",
                "tags": [
                    "Authentication"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "description": "User's email address"
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "User's password"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User authenticated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "userId": {
                                            "type": "string",
                                            "description": "The authenticated user's ID"
                                        },
                                        "token": {
                                            "type": "string",
                                            "description": "JWT token for accessing protected routes"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid email or password"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/books": {
            "get": {
                "summary": "Renvoie un tableau de tous les livres de la base de données.",
                "tags": [
                    "Books"
                ],
                "responses": {
                    "200": {
                        "description": "A list of books",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No books found"
                    }
                }
            },
            "post": {
                "summary": "Capture et enregistre l'image, analyse le livre transformé en chaîne de caractères, et l'enregistre dans la base de données en définissant correctement son ImageUrl.Initialise la note moyenne du livre à 0 et le rating avec un tableau vide. Remarquez que le corps de la demande initiale est vide ; lorsque Multer est ajouté, il renvoie une chaîne pour le corps de la demande en fonction des données soumises avec le fichier.",
                "tags": [
                    "Books"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "book": {
                                        "type": "string",
                                        "description": "Book data in JSON format"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "Cover image of the book"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/books/bestrating": {
            "get": {
                "summary": "Renvoie un tableau des 3 livres de la base de données ayant la meilleure note moyenne.",
                "tags": [
                    "Books"
                ],
                "responses": {
                    "200": {
                        "description": "A list of top 3 books",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "summary": "Get a book by ID",
                "tags": [
                    "Books"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "ID of the book to retrieve"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            },
            "put": {
                "summary": "Met à jour le livre avec l'_id fourni. Si une image est téléchargée, elle est capturée, et l’ImageUrl du livre est mise à jour. Si aucun fichier n'est fourni, les informations sur le livre se trouvent directement dans le corps de la requête (req.body.title, req.body.author, etc.). Si un fichier est fourni, le livre transformé en chaîne de caractères se trouve dans req.body.book. Notez que le corps de la demande initiale est vide ; lorsque Multer est ajouté, il renvoie une chaîne du corps de la demande basée sur les données soumises avec le fichier.",
                "tags": [
                    "Books"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "ID of the book to update"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "book": {
                                        "type": "string",
                                        "description": "Book data in JSON format"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "New cover image of the book"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Unauthorized request"
                    }
                }
            },
            "delete": {
                "summary": "Supprime le livre avec l'_id fourni ainsi que l’image associée.",
                "tags": [
                    "Books"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "ID of the book to delete"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book deleted successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/books/{id}/rating": {
            "post": {
                "summary": "Définit la note pour le user ID fourni. La note doit être comprise entre 0 et 5. L'ID de l'utilisateur et la note doivent être ajoutés au tableau rating afin de ne pas laisser un utilisateur noter deux fois le même livre. Il n’est pas possible de modifier une note. La note moyenne averageRating doit être tenue à jour, et le livre renvoyé en réponse de la requête",
                "tags": [
                    "Books"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "ID of the book to rate"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "userId": {
                                        "type": "string",
                                        "description": "ID of the user rating the book"
                                    },
                                    "rating": {
                                        "type": "integer",
                                        "description": "Rating value (0 to 5)",
                                        "minimum": 0,
                                        "maximum": 5
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rating added successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "User has already rated this book"
                    },
                    "404": {
                        "description": "Book not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "unique eMail address of the user",
                        "required": "true",
                        "unique": "true"
                    },
                    "password": {
                        "type": "string",
                        "required": "true"
                    }
                }
            },
            "Book": {
                "type": "object",
                "required": [
                    "userId",
                    "title",
                    "author",
                    "imageUrl",
                    "year",
                    "genre",
                    "ratings",
                    "averageRating"
                ],
                "properties": {
                    "userId": {
                        "type": "string",
                        "description": "ID of the user who added the book"
                    },
                    "title": {
                        "type": "string",
                        "description": "Title of the book"
                    },
                    "author": {
                        "type": "string",
                        "description": "Author of the book"
                    },
                    "imageUrl": {
                        "type": "string",
                        "description": "URL of the book's cover image"
                    },
                    "year": {
                        "type": "integer",
                        "description": "Publication year of the book"
                    },
                    "genre": {
                        "type": "string",
                        "description": "Genre of the book"
                    },
                    "ratings": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "userId": {
                                    "type": "string",
                                    "description": "ID of the user who rated the book"
                                },
                                "grade": {
                                    "type": "integer",
                                    "description": "Rating given by the user",
                                    "minimum": 0,
                                    "maximum": 5
                                }
                            }
                        },
                        "description": "List of ratings for the book"
                    },
                    "averageRating": {
                        "type": "number",
                        "format": "float",
                        "description": "Average rating of the book"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "securityDefinitions": {
            "Bearer": {
                "type": "JWT",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}